refactor(shelter): move sameMapBounds next to the other bounds utils
sameMapBounds is a bounds predicate, so it belongs alongside toMapBounds,
fromMapBounds and mapBoundsFromCenter in Map/utils rather than in
mapViewport.ts (which owns the center+zoom viewport helpers). Keeps the
utils folder cohesive and each helper's test colocated with its source.
