+++ libs/react/shelter/src/lib/components/Map/utils/index.ts
@@ -1,7 +1,8 @@
 export { fitMapToPlaceViewport } from './fitMapToPlaceViewport';
 export { fromMapBounds } from './fromMapBounds';
 export { mapBoundsFromCenter } from './mapBoundsFromCenter';
-export { mapViewportFromMap, sameMapBounds } from './mapViewport';
+export { mapViewportFromMap } from './mapViewport';
 export type { TMapViewport } from './mapViewport';
+export { sameMapBounds } from './sameMapBounds';
 export { toGoogleLatLng } from './toGoogleLatLng';
 export { toMapBounds } from './toMapBounds';
